Gracefully handling the case where Resource processes are not found a particular host.

+++ code/bin/kill-resource.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
+# Finds a process corresponding to Granules/Neptune resource and kills it.
+
+# First use JPS to locate the Java process.
 r_pid=`jps | grep 'ManagedResource' | cut -d' ' -f1`
-echo 'Resource PID: '${r_pid}
+
+# If JPS fails, use ps and grep
 if [ -z ${r_pid} ]; then 
     r_pid=`ps -ef | grep 'java' | grep ${USER} | grep 'granules' | cut -d' ' -f2`
-    echo '[PS] Resource PID: '${r_pid}
 fi
 
-`kill -9 ${r_pid}`
+if [ -z ${r_pid} ]; then
+    echo 'No Resource process found!'
+else
+    `kill -9 ${r_pid}`
+    echo 'Killed the Resource process! PID: '${r_pid}
+fi
